requirements.txt
supabase
python-dotenv
pandas
dateutil